chore: move runGeoJSONWorker into GeoJSONShapesDataProvider
