+++ packages/@tissuumaps-storage/src/geojson/GeoJSONShapesDataProvider.ts
@@ -12,8 +12,12 @@ import {
   type GeoJSONShapesDataSource,
   createDefaultGeoJSONShapesDataSource,
 } from "./GeoJSONShapesDataSource";
-import type { GeoJSONWorkerRequest } from "./geojson.worker";
-import { runGeoJSONWorker } from "./runGeoJSONWorker";
+import type {
+  GeoJSONWorkerRequest,
+  GeoJSONWorkerResponse,
+  GeoJSONWorkerResponseFor,
+} from "./geojson.worker";
+import GeoJSONWorker from "./geojson.worker?worker&inline";
 
 export class GeoJSONShapesDataProvider implements ShapesDataProvider<
   GeoJSONShapesDataSource,
@@ -99,9 +103,41 @@ export class GeoJSONShapesDataProvider implements ShapesDataProvider<
       throw new Error("A URL or workspace path is required to load data.");
     }
 
-    const { ids, names, geometry } = await runGeoJSONWorker(request, {
-      signal,
-    });
+    const { ids, names, geometry } = await GeoJSONShapesDataProvider._runWorker(
+      request,
+      { signal },
+    );
     return new GeoJSONShapesData(ids, names, geometry);
+  }
+
+  private static _runWorker<TRequest extends GeoJSONWorkerRequest>(
+    request: TRequest,
+    options?: { signal?: AbortSignal },
+  ): Promise<GeoJSONWorkerResponseFor<TRequest>> {
+    const { signal } = options ?? {};
+    signal?.throwIfAborted();
+    const worker = new GeoJSONWorker();
+    return new Promise((resolve, reject) => {
+      const onAbort = () => {
+        worker.terminate();
+        reject(signal!.reason as Error);
+      };
+      signal?.addEventListener("abort", onAbort, { once: true });
+      worker.onmessage = (event: MessageEvent<GeoJSONWorkerResponse>) => {
+        worker.terminate();
+        signal?.removeEventListener("abort", onAbort);
+        if ("error" in event.data) {
+          reject(new Error(event.data.error));
+        } else {
+          resolve(event.data as GeoJSONWorkerResponseFor<TRequest>);
+        }
+      };
+      worker.onerror = (event) => {
+        worker.terminate();
+        signal?.removeEventListener("abort", onAbort);
+        reject(new Error(event.message));
+      };
+      worker.postMessage(request);
+    });
   }
 }
